OOP has already been done prior to this increment. Renamed Printer package to Ui, and Database package to Storage
