+++ src/main/java/LogicHandlers/CommandHandlers/EventHandler.java
@@ -6,8 +6,8 @@ import Models.Event;
 import Models.TaskArray;
 
 import static LogicHandlers.Parsers.DateTimeParser.parseDateTimeFromString;
-import static Printers.BasicOutputPrinter.printBasicOutput;
-import static Printers.ErrorOutputPrinter.printErrorOutput;
+import static Ui.BasicOutputPrinter.printBasicOutput;
+import static Ui.ErrorOutputPrinter.printErrorOutput;
 
 /**
  * EventHandler handles all 'event' commands.
