+++ src/main/java/LogicHandlers/Parsers/DateTimeParser.java
@@ -6,8 +6,6 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-import static Printers.ErrorOutputPrinter.printErrorOutput;
-
 /**
  * This class parses datetimes in String and returns them in DateTime formats.
  */
